fix(attn): K/V dtype bug in three Metal attention kernels + parity tests
flash_attention.metal, attention_gqa.metal, and attention_splitk.metal
(compute kernel) bound K/V caches as `const device float*`, while the
caches are stored as f16 on the Rust side (encode_cast_f32_to_f16_offsets,
see the `// f16 bytes` comment in metal_lfm2.rs:2445). Reading f16 as f32
reinterprets two adjacent halves as one f32 — every KV load in these
kernels was garbage.

Impact:
- flash_attention auto-activates at seq_len > 4096 (metal_lfm2.rs:1626),
  so any context window past 4096 produced corrupt logits and silently
  unusable generation. This is the "14% slower than classic, needs
  rewrite" item in the memory note — not just slower, actually broken.
- attention_gqa is opt-in via WICK_ATTN=gqa (latent, but silently wrong
  whenever opted in).
- attention_splitk is opt-in via WICK_ATTN=splitk (same).
- Classic attention.metal and attention_prefill.metal bound correctly as
  half* — only these two were safe.

Fix: change K/V bindings to `half*` and cast to float on load (3 edits
per file).

Tests:
- wick/tests/metal_shaders_parity.rs: 3 new synthetic-input tests
  (test_classic_vs_{flash,gqa,splitk}_attention_synthetic) that run in
  CI — no model file needed. Construct Q/K/V buffers directly, run each
  kernel variant, compare outputs against classic (always-correct
  half* binding) with 1e-3 tolerance.
- wick/tests/attention_metal_parity.rs (new): 3 ignored end-to-end
  tests that reload the model with WICK_FLASH=1 / WICK_ATTN=gqa /
  WICK_ATTN=splitk and assert the greedy decode matches classic on a
  real 450M-Q4_0 model. Env-var mutations guarded by an RAII
  EnvVarGuard backed by a process-wide OnceLock<Mutex<()>>, restoring
  the prior value on Drop (panic-safe) and serializing concurrent
  tests.

All 16 metal_shaders_parity tests pass (13 pre-existing + 3 new). All
3 end-to-end parity tests pass post-fix (pre-fix they fail with
classic=6 coherent tokens vs variant=12 random tokens).

+++ wick/src/backend/shaders/attention_splitk.metal
@@ -24,10 +24,13 @@ struct SplitParams {
 // Phase A: per-split partial computation.
 // Dispatch: (n_heads × n_splits) threadgroups, 256 threads each.
 // Writes: partials_out[head, split, d], partials_max[head, split], partials_sum[head, split].
+// K/V caches are stored as f16 (see `encode_cast_f32_to_f16_offsets` in
+// metal_lfm2.rs and the `// f16 bytes` comment). Binding as `float*`
+// would reinterpret two adjacent halves as one f32 and produce garbage.
 kernel void attention_split_compute(
     const device float* q [[buffer(0)]],
-    const device float* k_cache [[buffer(1)]],
-    const device float* v_cache [[buffer(2)]],
+    const device half*  k_cache [[buffer(1)]],
+    const device half*  v_cache [[buffer(2)]],
     device float* partials_out [[buffer(3)]],    // [n_heads × n_splits × head_dim]
     device float* partials_max [[buffer(4)]],    // [n_heads × n_splits]
     device float* partials_sum [[buffer(5)]],    // [n_heads × n_splits]
@@ -91,7 +94,7 @@ kernel void attention_split_compute(
         float dot = 0.0f;
         uint k_base = t * kv_dim + kv_h_offset;
         for (uint d = 0u; d < head_dim; d++) {
-            dot += q_shared[d] * k_cache[k_base + d];
+            dot += q_shared[d] * float(k_cache[k_base + d]);
         }
         scores[i] = dot * scale;
     }
@@ -138,7 +141,7 @@ kernel void attention_split_compute(
     if (d < head_dim) {
         float val = 0.0f;
         for (uint ii = s_group; ii < t_len; ii += par) {
-            val += scores[ii] * v_cache[(t_start + ii) * kv_dim + kv_h_offset + d];
+            val += scores[ii] * float(v_cache[(t_start + ii) * kv_dim + kv_h_offset + d]);
         }
         for (uint offset = par >> 1; offset > 0u; offset >>= 1) {
             val += simd_shuffle_down(val, offset);
